SIMPLEX-123 Refactor.
